Move V1 API to the separate mod.
