+++ src/api/mod.rs
@@ -1,0 +1,2 @@
+pub mod server;
+pub mod v1;
